Rename estuary logo
